refactor(windows-ime): 恢复策略层迁入跨平台模块并纳入 CI 单测（#852 审查跟进）
- 新增 windows_ime_restore 模块：run_restore_flow/RestoreOutcome/重试等待
  迁出 windows_ime_session；sleep_restore_retry 增加 MultiThread runtime 守卫
- report_restore_step_results 提升为无门控顶层函数，测试可在任意平台执行
- restore_profile 参数收敛为 RestoreArgs 结构体（resolve_restore_args）
- 测试夹具 openless_snapshot_for_test 由生产常量派生，替换重复 GUID 字面量
- restore_session 消费 RestoreOutcome 补 debug 诊断日志
- backend-tests 纳入 windows_ime_profile/windows_ime_restore（tokio/winreg/features）
- 新增 9 个单测现已在 macOS/Linux cargo test --lib 与 Windows backend-tests 执行
